Move TargetToken to history parity module
